Create new class MapViewUtils but not sure if good idea
