+++ app/src/main/java/ch/epfl/sdp/ui/maps/MapViewUtils.kt
@@ -1,0 +1,53 @@
+package ch.epfl.sdp.ui.maps
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.mapbox.mapboxsdk.maps.MapView
+
+open class MapViewUtils : AppCompatActivity(){
+    private var mapView: MapView? = null
+
+    fun setMapView(mapView: MapView){
+        this.mapView = mapView
+    }
+
+    fun getMapView(): MapView? {
+        return this.mapView
+    }
+
+    // Override Activity lifecycle methods
+    public override fun onResume() {
+        super.onResume()
+        mapView!!.onResume()
+    }
+
+    override fun onStart() {
+        super.onStart()
+        mapView!!.onStart()
+    }
+
+    override fun onStop() {
+        super.onStop()
+        mapView!!.onStop()
+    }
+
+    public override fun onPause() {
+        super.onPause()
+        mapView!!.onPause()
+    }
+
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        mapView!!.onSaveInstanceState(outState)
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        mapView!!.onDestroy()
+    }
+
+    override fun onLowMemory() {
+        super.onLowMemory()
+        mapView!!.onLowMemory()
+    }
+}
